test-fridge/requirements.txt
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.6.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0




